feature(battle): スコアUIの作成とタイマーの背景配置
Issue:
- Closes #155
