+++ Assets/_Application/Mode/Textures.meta
