+++ Assets/_Application/Mode/Textures/matcha_score.png
