+++ Assets/_Application/Mode/Textures/strawberry_score.png
