+++ Assets/_Application/Mode/Uraty.Application.Mode.asmdef
@@ -4,7 +4,8 @@
     "references": [
         "GUID:aff03c096e22e514e830c9baff70dd4f",
         "GUID:b5c31c0c834ee7e4b980691b778889bf",
-        "GUID:251a6b1f97bff7349aead7bcd9069f8e"
+        "GUID:251a6b1f97bff7349aead7bcd9069f8e",
+        "GUID:6055be8ebefd69e48b49212b09b47b2f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
